# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "API support for a E-Commerce",
        "contact": {
            "email": "makinsuroju@gmail.com"
        },
        "license": {
            "name": "Apache License 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/password-reset/create": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Password reset request",
                "operationId": "78686995fdaea6c738b91ccb4d84a36e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "redirect_url"
                                ],
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "john-doe@mail.com"
                                    },
                                    "redirect_url": {
                                        "title": "redirect_url",
                                        "type": "string",
                                        "example": "http://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "We have e-mailed your password reset link!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "We can't find a user with that email address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/password-reset/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "Password reset with token",
                "operationId": "b1f447b4b3567651aae0b4c46a13bd90",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset with token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "title": "token",
                                            "type": "string",
                                            "example": "cVirkKfR6wg4NPbfjMAdEmzHacztBeP2twIPm6StoTDsvHUWymvcrlJD9cDB"
                                        },
                                        "email": {
                                            "title": "email",
                                            "type": "email",
                                            "example": "example@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Password reset with token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "This password reset token is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/password-reset/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "description": "Reset Password",
                "operationId": "30af651967c1ed4178a1f9fd41a2cb55",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "JohnDoe@mail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "title": "password_confirmation",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "token": {
                                        "title": "token",
                                        "type": "string",
                                        "example": "cVirkKfR6wg4NPbfjMAdEmzHacztBeP2twIPm6StoTDsvHUWymvcrlJD9cDB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "We can't find a user with that email address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Errors"
                    }
                }
            }
        },
        "/api/v1/password-change/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "description": "In application password change",
                "operationId": "1be4d2efbf5d0ff435f89d33cc33ddfb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "type": "email",
                                        "default": "JohnDoe@mail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "title": "password_confirmation",
                                        "type": "password",
                                        "default": "********",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "true"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "title": "success",
                                            "type": "boolean",
                                            "default": "false"
                                        },
                                        "message": {
                                            "title": "message",
                                            "type": "string",
                                            "default": "We can't find a user with that email address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Errors"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer",
                        "default": "1"
                    },
                    "first_name": {
                        "title": "first_name",
                        "type": "string",
                        "default": "John",
                        "minLength": 3
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string",
                        "default": "Doe",
                        "minLength": 3
                    },
                    "email": {
                        "title": "email",
                        "type": "email",
                        "default": "john-doe@mail.com"
                    },
                    "avatar": {
                        "title": "avatar",
                        "type": "string",
                        "default": "avatar.png"
                    },
                    "telephone": {
                        "title": "telephone",
                        "type": "telephone",
                        "default": "080 2566 4567"
                    },
                    "activation_token": {
                        "title": "activation_token",
                        "type": "string",
                        "example": "6wjKmRtjddXW79sKXsOrJ9ZnloOY5iuwVm9SlA9Le3xSR8ydD2ugljZwNhPZ"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}